Fix warnings

+++ source/numeric/VectorBuilder.cpp
@@ -171,7 +171,8 @@ SVInt VectorBuilder::finish() {
                 bits = SVInt::MAX_BITS;
             }
             if (sizeBits == 0) {
-                return SVInt(std::max(32, bits), literalBase, signFlag, hasUnknown, ArrayRef<logic_t>(digits.begin(), digits.count()));
+                return SVInt((uint16_t)std::max(32, bits), literalBase, signFlag, hasUnknown,
+                             ArrayRef<logic_t>(digits.begin(), digits.count()));
             } else {
                 // we should warn about overflow here, but the spec says it is valid and
                 // the literal gets truncated. Definitely a warning though.
@@ -180,7 +181,8 @@ SVInt VectorBuilder::finish() {
         }
     }
 
-    return SVInt(sizeBits ? sizeBits : 32, literalBase, signFlag, hasUnknown, ArrayRef<logic_t>(digits.begin(), digits.count()));
+    return SVInt(sizeBits ? sizeBits : 32, literalBase, signFlag, hasUnknown,
+                 ArrayRef<logic_t>(digits.begin(), digits.count()));
 }
 
 void VectorBuilder::addDigit(logic_t digit, int maxValue) {
